+++ springboot-data-jpa/src/main/java/com/learning/springboot/bootdatajpa/part01_data_jpa/LibraryBookRepository.java
@@ -1,0 +1,6 @@
+package com.learning.springboot.bootdatajpa.part01_data_jpa;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LibraryBookRepository extends JpaRepository<LibraryBook, Long> {
+}